% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatemplate.R
\docType{data}
\name{df_temp}
\alias{df_temp}
\title{Empty template for the SLEEPR redcap database}
\format{
A data frame with 0 rows and 519 variables:
}
\source{
George Fulk
}
\usage{
df_temp
}
\description{
A dataset containing no entries
}
\keyword{datasets}
