% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeprutil.R
\docType{package}
\name{sleeprutil}
\alias{sleeprutil}
\title{sleeprutil: Utilities for NIH (SLEEPR) grant}
\description{
The sleeprutil package provides data support functionality for the SLEEPR grant
}
\section{sleeprutil functions}{

The foo functions ...
}

