% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_oxi_data.R
\name{template_oxi_data}
\alias{template_oxi_data}
\title{template_oxi_data}
\usage{
template_oxi_data(testing = FALSE)
}
\arguments{
\item{testing}{logical (FALSE) set to true only for testing}
}
\value{
Nothing.  Runs for side effect (creates a file).
In test mode, returns dataframe of output that would written to csv
}
\description{
template_oxi_data
}
