% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_pal_data.R
\name{template_pal_data}
\alias{template_pal_data}
\title{template_pal_data}
\usage{
template_pal_data()
}
\value{
Nothing.  Runs for side effect (creates a file)
}
\description{
Allows user to interactively select a .csv file created from PAL device and
outputs a file of similar name, but inserted into the REDCAP template.
}
\examples{

}
